refactor code for #1 - #6
